+++ Source/DataSources/ModelGraphics.js
@@ -50,6 +50,8 @@ define([
      * @param {Property} [options.shadows=ShadowMode.ENABLED] An enum Property specifying whether the model casts or receives shadows from each light source.
      * @param {Property} [options.heightReference=HeightReference.NONE] A Property specifying what the height is relative to.
      * @param {Property} [options.distanceDisplayCondition] A Property specifying at what distance from the camera that this model will be displayed.
+     * @param {Property} [options.silhouetteColor=Color.RED] A Property specifying the {@link Color} of the silhouette.
+     * @param {Property} [options.silhouetteSize=0.0] A numeric Property specifying the size of the silhouette in pixels.
      * @param {Property} [options.color=Color.WHITE] A Property specifying the {@link Color} that blends with the model's rendered color.
      * @param {Property} [options.colorBlendMode=ColorBlendMode.HIGHLIGHT] An enum Property specifying how the color blends with the model.
      * @param {Property} [options.colorBlendAmount=0.5] A numeric Property specifying the color strength when the <code>colorBlendMode</code> is <code>MIX</code>. A value of 0.0 results in the model's rendered color while a value of 1.0 results in a solid color, with any value in-between resulting in a mix of the two.
@@ -80,6 +82,10 @@ define([
         this._heightReferenceSubscription = undefined;
         this._distanceDisplayCondition = undefined;
         this._distanceDisplayConditionSubscription = undefined;
+        this._silhouetteColor = undefined;
+        this._silhouetteColorSubscription = undefined;
+        this._silhouetteSize = undefined;
+        this._silhouetteSizeSubscription = undefined;
         this._color = undefined;
         this._colorSubscription = undefined;
         this._colorBlendMode = undefined;
@@ -198,6 +204,22 @@ define([
         distanceDisplayCondition : createPropertyDescriptor('distanceDisplayCondition'),
 
         /**
+         * Gets or sets the Property specifying the {@link Color} of the silhouette.
+         * @memberof ModelGraphics.prototype
+         * @type {Property}
+         * @default Color.RED
+         */
+        silhouetteColor: createPropertyDescriptor('silhouetteColor'),
+
+        /**
+         * Gets or sets the numeric Property specifying the size of the silhouette in pixels.
+         * @memberof ModelGraphics.prototype
+         * @type {Property}
+         * @default 0.0
+         */
+        silhouetteSize : createPropertyDescriptor('silhouetteSize'),
+
+        /**
          * Gets or sets the Property specifying the {@link Color} that blends with the model's rendered color.
          * @memberof ModelGraphics.prototype
          * @type {Property}
@@ -245,6 +267,8 @@ define([
         result.nodeTransformations = this.nodeTransformations;
         result.heightReference = this._heightReference;
         result.distanceDisplayCondition = this.distanceDisplayCondition;
+        result.silhouetteColor = this.silhouetteColor;
+        result.silhouetteSize = this.silhouetteSize;
         result.color = this.color;
         result.colorBlendMode = this.colorBlendMode;
         result.colorBlendAmount = this.colorBlendAmount;
@@ -275,6 +299,8 @@ define([
         this.runAnimations = defaultValue(this.runAnimations, source.runAnimations);
         this.heightReference = defaultValue(this.heightReference, source.heightReference);
         this.distanceDisplayCondition = defaultValue(this.distanceDisplayCondition, source.distanceDisplayCondition);
+        this.silhouetteColor = defaultValue(this.silhouetteColor, source.silhouetteColor);
+        this.silhouetteSize = defaultValue(this.silhouetteSize, source.silhouetteSize);
         this.color = defaultValue(this.color, source.color);
         this.colorBlendMode = defaultValue(this.colorBlendMode, source.colorBlendMode);
         this.colorBlendAmount = defaultValue(this.colorBlendAmount, source.colorBlendAmount);