+++ Specs/createScene.js
@@ -40,6 +40,7 @@ define([
         var contextOptions = options.contextOptions;
         contextOptions.webgl = defaultValue(contextOptions.webgl, {});
         contextOptions.webgl.antialias = defaultValue(contextOptions.webgl.antialias, false);
+        contextOptions.webgl.stencil = defaultValue(contextOptions.webgl.stencil, true);
         contextOptions.webglStub = !!window.webglStub;
 
         var scene = new Scene(options);